Refactored error module to include new server module
