+++ intmodules/__init__.py
@@ -1,4 +1,5 @@
-from intmodules.log_module import config_logger, logger
+from intmodules.log import config_logger
+from intmodules.log import logger
 from intmodules.preprocessor import preprocess_data
 from intmodules.json_validation import validate_request
 from intmodules.server_handler import MPBRequestHandler
